x/vm/types/message_cron_test.go
package types

import (
	"testing"

	"github.com/VestaProtocol/vesta/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCron_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCron
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCron{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCron{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
